# wikify-web/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Wikify Web API",
    "description": "AI-powered repository documentation and chat system",
    "contact": {
      "name": "Wikify Team",
      "email": "support@wikify.dev"
    },
    "license": {
      "name": "MIT OR Apache-2.0"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    },
    {
      "url": "https://api.wikify.dev",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/wiki/generate": {
      "post": {
        "tags": [
          "Wiki"
        ],
        "summary": "Generate wiki documentation",
        "description": "Generate comprehensive wiki documentation for a repository",
        "operationId": "generate_wiki",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateWikiRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Wiki generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateWikiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Session not found"
          },
          "500": {
            "description": "Wiki generation failed"
          }
        }
      }
    },
    "/api/wiki/{repository_id}": {
      "get": {
        "tags": [
          "Wiki"
        ],
        "summary": "Get generated wiki",
        "description": "Retrieve the generated wiki documentation for a repository. No authentication required.",
        "operationId": "get_wiki",
        "parameters": [
          {
            "name": "repository_id",
            "in": "path",
            "description": "Repository ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Wiki retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WikiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Repository not found"
          },
          "500": {
            "description": "Failed to generate wiki"
          }
        }
      }
    },
    "/api/wiki/{repository_id}/export": {
      "post": {
        "tags": [
          "Wiki"
        ],
        "summary": "Export wiki",
        "description": "Export generated wiki in various formats (not yet implemented)",
        "operationId": "export_wiki",
        "parameters": [
          {
            "name": "repository_id",
            "in": "path",
            "description": "Repository ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Wiki exported successfully"
          },
          "501": {
            "description": "Not implemented"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "GenerateWikiRequest": {
        "type": "object",
        "description": "Wiki generation request",
        "required": [
          "repository_id",
          "config"
        ],
        "properties": {
          "config": {
            "$ref": "#/components/schemas/WikiGenerationConfig"
          },
          "repository_id": {
            "type": "string",
            "example": "uuid-string"
          }
        }
      },
      "GenerateWikiResponse": {
        "type": "object",
        "description": "Wiki generation response",
        "required": [
          "wiki_id",
          "status",
          "pages_count",
          "sections_count"
        ],
        "properties": {
          "pages_count": {
            "type": "integer",
            "example": 25,
            "minimum": 0
          },
          "sections_count": {
            "type": "integer",
            "example": 8,
            "minimum": 0
          },
          "status": {
            "type": "string",
            "example": "success"
          },
          "wiki_id": {
            "type": "string",
            "example": "uuid-string"
          }
        }
      },
...
      "WikiGenerationConfig": {
        "type": "object",
        "description": "Wiki generation configuration",
        "properties": {
          "comprehensive_view": {
            "type": [
              "boolean",
              "null"
            ],
            "example": false
          },
          "include_diagrams": {
            "type": [
              "boolean",
              "null"
            ],
            "example": true
          },
          "language": {
            "type": [
              "string",
              "null"
            ],
            "example": "en"
          },
          "max_pages": {
            "type": [
              "integer",
              "null"
            ],
            "example": 50,
            "minimum": 0
          }
        }
      },
      "WikiPageResponse": {
        "type": "object",
        "description": "Wiki page response structure",
        "required": [
          "id",
          "title",
          "content",
          "description",
          "importance",
          "file_paths",
          "related_pages",
          "tags",
          "reading_time",
          "generated_at",
          "source_documents"
        ],
        "properties": {
          "content": {
            "type": "string",
            "example": "# Main Documentation\n\nThis is the main documentation..."
          },
          "description": {
            "type": "string",
            "example": "Main documentation page"
          },
          "file_paths": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "generated_at": {
            "type": "string",
            "example": "2024-01-01T00:00:00Z"
          },
          "id": {
            "type": "string",
            "example": "main"
          },
          "importance": {
            "type": "string",
            "example": "Critical"
          },
          "reading_time": {
            "type": "integer",
            "example": 5,
            "minimum": 0
          },
          "related_pages": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "source_documents": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "title": {
            "type": "string",
            "example": "Main Documentation"
          }
        }
      },
      "WikiResponse": {
        "type": "object",
        "description": "Wiki response structure",
        "required": [
          "id",
          "title",
          "description",
          "pages",
          "sections"
        ],
        "properties": {
          "description": {
            "type": "string",
            "example": "Generated wiki for repository"
          },
          "id": {
            "type": "string",
            "example": "repo_123"
          },
          "pages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WikiPageResponse"
            }
          },
          "sections": {
            "type": "array",
            "items": {}
          },
          "title": {
            "type": "string",
            "example": "Repository Wiki"
          }
        }
      }
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
...
}